pyproject.toml
[tool.poetry]
name = "goit-web-module6"
version = "0.1.0"
description = ""
authors = ["Nataliia Tiutiunnyk <nat.tiutiunnyk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
faker = "^22.0.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
